+++ lib/LibCraftingProfessions-1.0/CHANGES.md
@@ -1,3 +1,7 @@
+### Revision 10 (2025/05/30)
+- Add profession frame SHOW/CLOSE events
+- Support all modern major 1.12 profession frame replacement addons with all their quirks and little details ([AdvancedTradeSkillWindow](https://github.com/laytya/AdvancedTradeSkillWindow-vanilla), [AdvancedTradeSkillWindow2](https://github.com/Shellyoung/AdvancedTradeSkillWindow2), [Artisan](https://github.com/Otari98/Artisan))
+
 ### Revision 5 (2025/05/03)
 - Support Turtle WoW Jewelcrafting profession
 - Add esMX and ptBR locales